Refactor: Wrap row indices and table buffers for code safety
New classes BufferRowIndex and ViewRowIndex are designed to make sure the two (buffer/view row indices) are never conflated in code. Constructors from int are explicit so that they always need to be spelled out.

On the other side of the equation (the place where most of these indices are actually used), the buffers in tables are now also wrapped to keep the places where indices have to be unwrapped as compact and centralized as possible, because that is where mistakes can still happen. Unfortunately, this encapsulation does not work perfectly because buffer indices sometimes also have to be unwrapped to create a QModelIndex or make calls to beginRemoveRows() etc.

Closes #166