Refactor: Handling of different kinds of composite columns
New nomenclature and implementation for composite tables and columns: There are now normal column, export-only columns and filter-only columns.
Normal and filter-only columns are buffered and their models stored in the same list. Export-only columns are unbuffered and stored in a separate list, along with the index of the normal column they go before.

Add some new helper functions for the extended functionality.

For #175