Runtime type checking for ItemIDs
Every ItemID is assigned an item type (like ascent, trip...) at creation. When attempting to extract the id value somewhere else, the expected type has to be provided separately and is asserted to match the stored type.

Type checking at compile time would be vastly preferrable but seems to be unfeasable because of template restrictions in Qt and complications from multiple inheritance.

Missing documentation changes and more intensive testing.