Change how versions are specified in settings update code
Instead of giving the version where the changes are introduced, which might not be known at development time (will the next release be 1.3.1 or 1.4.0?), specify the version before that, i.e. the most recent release at development time. This version and any version older than it are considered to need the corresponding changes applied.